move VARARGS_DELIM into pycbc/__init__.py
